Monday's work
Def a lot of things, mostly related to the services pages. Put it all
in flexbox and left the asides completely in their own columns (as
opposed to bleeding over into the main content). Also made small
changes to the front page, including the “Latest News” caption and
shrinking the boxes (though, that may have been instituted in the last
update).